feat(types): move types to index.d.ts
