--- src/core/aliasesDictionary.d.ts
@@ -1,16 +1,0 @@
-declare type AliasesDictionary = {
-  bot: "bottom";
-  col: "column";
-  dir: "direction";
-  fx: "flex";
-  lt: "left";
-  rt: "right";
-  bg: "background";
-  txt: "text";
-  jf: "justify";
-  pd: "padding";
-  wd: "width";
-  hg: "height";
-}
-
-export default AliasesDictionary;
