--- src/core/pathTransform.d.ts
@@ -1,9 +1,0 @@
-export let separator: string;
-export function setSeparator(sp: string): void;
-
-declare function pathTransform(
-  path: string,
-  getConstant: Function
-): object;
-
-export default pathTransform;
